separated custom exam sheet funtions into CustomExamSheetComponenet.tsx in components folder but not styling into app.css yet
